Add the HTML files for Unit 2.
The as-downloaded sample1 and sample2 html files are included in the
directory, and a new HTML file, created myself are added.

The HTML file I have created myself was informed by my reading on MDN,
one of the best resources on web development (in my opinion). Individual
comments are included (if necessary) to indicate others' ideas, code, or
content.

Included in the HTML file I have created is a quotation. Using the HTML
blockquote element, with a cite attribute, an a cite tag following is
sufficient, in my opinion, for citation and attribution. Given the
course materials on The Landing indicate that a simple hyperlink is
often enough, this is above and beyond.