Git cleanup
There were many files which were not needed. The top-level directory
structure did not make sense, so I am moving the actual website from a
subdirectory to the top-level directory. The next commit will perform
the move operation.

This commit is mostly deleting the unnecessary garbage, but also adds a
couple new files. This commit should be ammended.